New buffer navigation plugin added!
Also added new keymaps such as gd and gD for navigating
definitions/declarations and others. Also added linters for html and css
but for some reasons I can't seem to find a way to get the html linter
working or I think it is not supposed to be used that way?